% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say "Hello" to arg}
\usage{
hello(name)
}
\arguments{
\item{whom}{the function says "Hello" to}
}
\value{
chr
}
\description{
Say "Hello" to arg
}
\examples{
hello("Bob")
hello("My friend")
}
